+++ ldm/invoke/config/invokeai_update.py
@@ -42,7 +42,13 @@ def invokeai_is_running()->bool:
         except psutil.AccessDenied:
             continue
     return False
-        
+
+def do_post_install():
+    '''
+    Run postinstallation script.
+    '''
+    from ldm.invoke.config.post_install.py import post_install
+    post_install()
     
 def welcome(versions: dict):
     
@@ -107,6 +113,7 @@ def main():
         print(f':heavy_check_mark: Upgrade successful')
     else:
         print(f':exclamation: [bold red]Upgrade failed[/red bold]')
+    do_post_install()
     
 if __name__ == "__main__":
     try:
